+++ SplunkInteractions/Tests/SplunkInteractionsTests/Support/TestInteractionDestination.swift
@@ -20,10 +20,16 @@ import Foundation
 @testable import SplunkInteractions
 
 final class TestInteractionDestination: SplunkInteractionsDestination {
+
     var didReceiveInteractionCallCount = 0
     var actionName: String?
 
-    func send(actionName: String, elementId _: String?, time _: Date) {
+    func send(
+        actionName: String,
+        elementId _: String?,
+        xpath _: String?,
+        time _: Date
+    ) {
         self.actionName = actionName
         didReceiveInteractionCallCount += 1
     }