+++ SplunkOpenTelemetryBackgroundExporter/Tests/SplunkOpenTelemetryBackgroundExporterTests/Test Models/OTLPAdapterGroupingTests.swift
@@ -182,7 +182,8 @@ struct OTLPAdapterGroupingTests {
                 return value
             }
         )
-        let expectedScopeAttributeValues: Set = ["value-a", "value-b"]
+        let expectedScopeAttributeValues = Set(["value-a", "value-b"])
+
         #expect(scopeAttributeValues == expectedScopeAttributeValues)
     }
 
@@ -245,7 +246,12 @@ struct OTLPAdapterGroupingTests {
                 return value
             }
         )
+<<<<<<< HEAD
         let expectedScopeAttributeValues: Set = ["value-a", "value-b"]
+=======
+        let expectedScopeAttributeValues = Set(["value-a", "value-b"])
+
+>>>>>>> origin/develop
         #expect(scopeAttributeValues == expectedScopeAttributeValues)
     }