+++ CHANGELOG.md
@@ -7,6 +7,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+* The SDK now uses static binary dependencies to enhance integration in complex deployment scenarios.
+
 ## [2.1.0] - 2026-02-23
 
 ### Added