+++ CHANGELOG.md
@@ -7,7 +7,13 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
 * Added support for adjust Session replay sampling rate.
+
+### Changed
+
+* The SDK now uses static binary dependencies to enhance integration in complex deployment scenarios.
 
 ## [2.1.0] - 2026-02-23