+++ CHANGELOG.md
@@ -9,7 +9,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-* Added support for adjust Session replay sampling rate.
+* Added support for adjusting the Session Replay sampling rate.
 
 ### Changed